requirements.txt
Flask
python-dotenv
firebase-admin
python-magic-bin ; platform_system == "Windows"
PyMuPDF
pdfplumber
